refactor(frontend): move VCSIcon into Release folder and remove VCS folder
VCSIcon is only used within the Release components, so colocate it
there and remove the now-empty VCS folder.
