+++ frontend/src/components/Release/VCSIcon.vue
