--- frontend/src/components/VCS/index.ts
@@ -1,3 +1,0 @@
-import VCSIcon from "./VCSIcon.vue";
-
-export { VCSIcon };
